update int N to double

+++ Confliction/Confliction.java
@@ -9,7 +9,8 @@ import java.util.Set;
 
     public class DoublePQ {
         private DoubleCar [] pq;                    // store items at indices 1 to N
-        private int N;                       // number of items on priority queue
+        
+        private double N;                       // number of items on priority queue
         private int P1;//postion in pindexpq
         private int P2;//positon in mindexpq
         private HashMap<String,Integer> pindexpq=new HashMap<String,Integer>();//store position ranked by price in PQ
